- Minor change in plugins:
The field <<url_regex>> has been replaced by the <<site>> field, which specifies the domain of the site.

For example:

	url_regex => qr/example\.net/,

became:

	site => 'example.net',

In easy cases, this change is backwards compatible; clyrics tries to convert the regular expression into a string, otherwise it will display an warning message and the broken plugin will be ignored.

- Major optimization in the search engine: we now search for lyrics only in the websites specified in plugins. This change made clyrics much faster and more accurate.
- Version bump to version: 0.06